CodeBreaker is working correctly as far as I can tell.
Had difficulty with determining if values were correct but not in the
correct place. Current method in place using a stack appears to be
working correctly.

First values that are not immediately correct and in the correct place
are pushed to a stack. The stack is then trimmed so that no duplicate
values are present. The values in the guess that are not correct and in
the correct place are then compared to the value on the trimmed stack.
The correct value counter is incremented if there is a match between the
values on the stack and the values in the guess.